# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Salmon Wallet Discovery API",
    "version": "1.0.0",
    "description": "Read-only discovery for Salmon Wallet's canonical public resources. The website does not expose a transactional, custodial, or wallet-management API. Breaking changes are released under a new URL major version such as /api/v2. A version scheduled for retirement will return RFC 9745 Deprecation and RFC 8594 Sunset headers at least 180 days before removal."
  },
  "servers": [{ "url": "https://salmonwallet.io", "description": "Canonical public website" }],
  "paths": {
    "/api/v1/discovery": {
      "get": {
        "operationId": "getDiscovery",
        "summary": "Get canonical Salmon Wallet resources",
        "description": "Returns stable links for product access and agent-oriented documentation. This operation never signs transactions or accesses wallet data.",
        "responses": {
          "200": {
            "description": "Canonical public resources",
            "headers": {
              "API-Version": { "description": "Major API version.", "schema": { "type": "string", "const": "1" } },
              "Deprecation": { "$ref": "#/components/headers/Deprecation" },
              "Sunset": { "$ref": "#/components/headers/Sunset" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Discovery" } } }
          },
          "406": { "$ref": "#/components/responses/NotAcceptable" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "headers": {
      "Deprecation": {
        "description": "Present only after this version is formally deprecated; value is an RFC 9745 date.",
        "schema": { "type": "string", "example": "@1798761600" }
      },
      "Sunset": {
        "description": "Present only for a retiring version; value is an HTTP-date at least 180 days after deprecation is announced.",
        "schema": { "type": "string", "format": "http-date", "example": "Wed, 30 Jun 2027 23:59:59 GMT" }
      }
    },
    "schemas": {
      "Discovery": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "description", "version", "website", "agentIndex", "openapi", "source", "products"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "version": { "type": "string", "const": "1" },
          "website": { "type": "string", "format": "uri" },
          "agentIndex": { "type": "string", "format": "uri" },
          "openapi": { "type": "string", "format": "uri" },
          "source": { "type": "string", "format": "uri" },
          "products": {
            "type": "object",
            "additionalProperties": false,
            "required": ["webWallet", "browserExtension", "android"],
            "properties": {
              "webWallet": { "type": "string", "format": "uri" },
              "browserExtension": { "type": "string", "format": "uri" },
              "android": { "type": "string", "format": "uri" }
            }
          }
        }
      },
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details with a stable machine-readable Salmon error code.",
        "required": ["type", "title", "status", "detail", "code"],
        "properties": {
          "type": { "type": "string", "format": "uri-reference" },
          "title": { "type": "string" },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string" },
          "instance": { "type": "string", "format": "uri-reference" },
          "code": { "type": "string", "description": "Stable machine-readable error identifier." }
        }
      }
    },
    "responses": {
      "NotAcceptable": {
        "description": "The requested representation is not available.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "InternalError": {
        "description": "Unexpected server error.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      }
    }
  },
  "externalDocs": { "description": "Salmon Wallet agent guidance", "url": "https://salmonwallet.io/llms.txt" }
}